Python/FastAPI/streamVideo_acLEGADA/RaspiCam/requirements.txt
fastapi==0.61.1
uvicorn==0.11.8
aiofiles==0.5.0
cookies==2.2.1
fastapi==0.61.1
gpiozero==1.5.1
imutils==0.5.3
isort==4.3.4
Jinja2==2.10
numpy==1.16.2
opencv-python==3.4.6.27
picamera==1.13
pyserial==3.4
requests==2.21.0
requests-oauthlib==1.0.0
responses==0.9.0
RPi.GPIO==0.7.0
starlette==0.13.6
urllib3==1.24.1
uvicorn==0.11.8
virtualenv==20.0.31
websockets==8.1